DockerFR/requirements.txt
fastapi
uvicorn[standard]
numpy
pillow
opencv-python
torch
torchvision
insightface
deepface
